move kv operations to erocksdb_kv.(cc|h)
